siggen/cache: Optionally allow adding siggen hash data to the bitbake cache
Being able to track siggen hash construction data can be useful for cache
debugging. For now, add an extra cache class which contains this information.
It can be enabled in the same way as the hob data cache through a feature flag
to cooker. This allows us to experiment with the data without carrying larger
patches around and ultimately may allow use to have a hash mismatch debugging
mode that is more easily enabled.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>