cooker: Avoid sideeffects for autorev from getAllKeysWithFlags
If we expand the variable AUTOREV in OE-Core, it triggers side effects in the
fetcher. The situation isn't ideal and needs improvement but this breaks
and is blocking enabling BB_DEFAULT_EVENTLOG.

Hack around the issue for now so we unblock things until we can work out
a better plan for how to improve AUTOREV support.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>